+++ src/devices/AlexaDevice.java
@@ -4,6 +4,8 @@ public class AlexaDevice {
     private AlexaSpeaker speaker = new AlexaSpeaker();
     private AlexaMic mic = new AlexaMic();
 
+    private AlexaDevice(){}
+
     public AlexaSpeaker getSpeaker() {
         return speaker;
     }
@@ -26,5 +28,35 @@ public class AlexaDevice {
                 "speaker=" + speaker.toString() +
                 ", mic=" + mic.toString() +
                 '}';
+    }
+
+    public static class AlexaBuilder {
+        private AlexaDevice alexa;
+
+        public AlexaBuilder() {
+            alexa = new AlexaDevice();
+        }
+
+        public AlexaSpeaker.AlexaSpeakerBuilder defineSpeaker() {
+            return new AlexaSpeaker.AlexaSpeakerBuilder(this);
+        }
+
+        public AlexaMic.AlexaMicBuilder defineMic(){
+            return new AlexaMic.AlexaMicBuilder(this);
+        }
+
+        public AlexaBuilder saveSpeaker(AlexaSpeaker speaker) {
+            alexa.setSpeaker(speaker);
+            return this;
+        }
+
+        public AlexaBuilder saveMic(AlexaMic mic) {
+            alexa.setMic(mic);
+            return this;
+        }
+
+        public AlexaDevice buildAlexa() {
+            return alexa;
+        }
     }
 }
