+++ src/devices/AlexaSpeaker.java
@@ -1,7 +1,5 @@
 package devices;
 
-import builder.AlexaBuilder;
-
 public class AlexaSpeaker {
     private Integer volume;
     private Integer minHz;
@@ -18,9 +16,9 @@ public class AlexaSpeaker {
 
     public static class AlexaSpeakerBuilder {
         private AlexaSpeaker speaker = new AlexaSpeaker();
-        private AlexaBuilder parentBuilder;
+        private AlexaDevice.AlexaBuilder parentBuilder;
 
-        public AlexaSpeakerBuilder(AlexaBuilder parentBuilder) {
+        public AlexaSpeakerBuilder(AlexaDevice.AlexaBuilder parentBuilder) {
             this.parentBuilder = parentBuilder;
         }
 
@@ -34,7 +32,7 @@ public class AlexaSpeaker {
             return this;
         }
 
-        public AlexaBuilder buildSpeaker() {
+        public AlexaDevice.AlexaBuilder buildSpeaker() {
             parentBuilder.saveSpeaker(speaker);
             return parentBuilder;
         }
